+++ rcsb/app/config/config.yml
@@ -14,7 +14,7 @@ configuration:
   SESSION_DIR_PATH: rcsb/app/tests-file/data/sessions
   SHARED_LOCK_PATH: rcsb/app/tests-file/data/shared-locks  # default or ternary lock only
   LOCK_TRANSACTIONS: True
-  LOCK_TYPE: sqlite # soft, ternary, or redis (need not agree with kv mode)
+  LOCK_TYPE: soft # soft, ternary, or redis (need not agree with kv mode)
   LOCK_TIMEOUT: 60
   # database parameters
   KV_MODE: sqlite # redis or sqlite, redis for multiple machines or containers, sqlite possible for one machine or container only
