+++ rcsb/app/file/KvConnection.py
@@ -9,18 +9,19 @@ from fastapi.exceptions import HTTPException
 
 
 class KvConnection(object):
-    def __init__(self, filepath, sessionTable, logTable, lockTable):
+    def __init__(self, filepath, sessionTable, mapTable, lockTable):
         self.filePath = filepath
         self.sessionTable = sessionTable
-        self.logTable = logTable
+        self.mapTable = mapTable
         self.lockTable = lockTable
         try:
+            # string interpolation for table names but not for data
             with self.getConnection() as connection:
                 connection.cursor().execute(
                     f"CREATE TABLE IF NOT EXISTS {self.sessionTable} (key, val)"
                 )
                 connection.cursor().execute(
-                    f"CREATE TABLE IF NOT EXISTS {self.logTable} (key,val)"
+                    f"CREATE TABLE IF NOT EXISTS {self.mapTable} (key,val)"
                 )
                 connection.cursor().execute(
                     f"CREATE TABLE IF NOT EXISTS {self.lockTable} (key,val)"
@@ -38,11 +39,17 @@ class KvConnection(object):
         res = None
         try:
             with self.getConnection() as connection:
+                params = (key,)
                 res = (
                     connection.cursor()
-                    .execute(f"SELECT val FROM {table} WHERE key = '{key}'")
+                    .execute(f"SELECT val FROM {table} " + "WHERE key = ?", params)
                     .fetchone()[0]
                 )
+                # res = (
+                #     connection.cursor()
+                #     .execute(f"SELECT val FROM {table} WHERE key = '{key}'")
+                #     .fetchone()[0]
+                # )
         except Exception:
             pass
         return res
@@ -62,20 +69,34 @@ class KvConnection(object):
     def set(self, key, val, table):
         try:
             with self.getConnection() as connection:
+                params = (key,)
                 res = (
                     connection.cursor()
-                    .execute(f"SELECT val FROM {table} WHERE key = '{key}'")
+                    .execute(f"SELECT val FROM {table} " + "WHERE key = ?", params)
                     .fetchone()
                 )
+                # res = (
+                #     connection.cursor()
+                #     .execute(f"SELECT val FROM {table} WHERE key = '{key}'")
+                #     .fetchone()
+                # )
                 if res is None:
+                    params = (key, val,)
                     res = connection.cursor().execute(
-                        f"INSERT INTO {table} VALUES ('{key}', \"{val}\")"
+                        f"INSERT INTO {table} " + "VALUES (?, ?)", params
                     )
+                    # res = connection.cursor().execute(
+                    #     f"INSERT INTO {table} VALUES ('{key}', \"{val}\")"
+                    # )
                     connection.commit()
                 else:
+                    params = (val, key,)
                     res = connection.cursor().execute(
-                        f"UPDATE {table} SET val = \"{val}\" WHERE key = '{key}'"
+                        f"UPDATE {table} " + "SET val = ? WHERE key = ?", params
                     )
+                    # res = connection.cursor().execute(
+                    #     f"UPDATE {table} SET val = \"{val}\" WHERE key = '{key}'"
+                    # )
                     connection.commit()
         except Exception as exc:
             logging.warning(
@@ -90,7 +111,9 @@ class KvConnection(object):
     def clear(self, key, table):
         try:
             with self.getConnection() as connection:
-                connection.cursor().execute(f"DELETE FROM {table} WHERE key = '{key}'")
+                params = (key,)
+                connection.cursor().execute(f"DELETE FROM {table} " + "WHERE key = ?", params)
+                # connection.cursor().execute(f"DELETE FROM {table} WHERE key = '{key}'")
                 connection.commit()
         except Exception as exc:
             logging.warning(
@@ -100,7 +123,9 @@ class KvConnection(object):
     def deleteRowWithKey(self, key, table):
         try:
             with self.getConnection() as connection:
-                connection.cursor().execute(f"DELETE FROM {table} WHERE key = '{key}'")
+                params = (key,)
+                connection.cursor().execute(f"DELETE FROM {table} " + "WHERE key = ?", params)
+                # connection.cursor().execute(f"DELETE FROM {table} WHERE key = '{key}'")
                 connection.commit()
         except Exception as exc:
             logging.warning("error in Kv delete from %s, %s %s", table, type(exc), exc)
@@ -108,7 +133,9 @@ class KvConnection(object):
     def deleteRowWithVal(self, val, table):
         try:
             with self.getConnection() as connection:
-                connection.cursor().execute(f"DELETE FROM {table} WHERE val = '{val}'")
+                params = (val,)
+                connection.cursor().execute(f"DELETE FROM {table} " + "WHERE val = ?", params)
+                # connection.cursor().execute(f"DELETE FROM {table} WHERE val = '{val}'")
                 connection.commit()
         except Exception as exc:
             logging.warning("error in Kv delete from %s, %s %s", table, type(exc), exc)
