+++ rcsb/app/tests-file/testKvConnection.py
@@ -15,9 +15,9 @@ class KvConnectionTest(unittest.TestCase):
         cP = ConfigProvider()
         filePath = cP.get("KV_FILE_PATH")
         sessionTable = cP.get("KV_SESSION_TABLE_NAME")
-        logTable = cP.get("KV_LOG_TABLE_NAME")
+        mapTable = cP.get("KV_MAP_TABLE_NAME")
         lockTable = cP.get("KV_LOCK_TABLE_NAME")
-        kV = KvConnection(filePath, sessionTable, logTable, lockTable)
+        kV = KvConnection(filePath, sessionTable, mapTable, lockTable)
         self.assertTrue(kV is not None)
         connection = kV.getConnection()
         self.assertTrue(connection is not None)
